v1.02 finished and released!
v1.02 comes with a much better way to set animation interpolators. The
entire SprintInterpolators was removed, and instead each Interpolator
is it’s own class. You can also pass in an easing type (ease in, out,
in + out) to an interpolator. This makes code much shorter and easier
to use! In addition, the documentation and comments were cleaned up,
and javadocs were generated.